Various fixes:   * QTantimony now properly finds modules to delete.   * Added local conversion factors.  This included re-writing the code to compare formulas to see if they had changed.   * Added USE_COMP in more places where it was needed.   * Code that synchronizes variables from SBML completely overhauled to account for rules and initial assignments also being synchronized (or not).   * Variables that are unset now write out that they are unset.  This turns out to be helpful when you want to delete information, and is also nice in a 'hey, here's what you didn't set yet' kind of way.     * This includes rate rules.

+++ antimony/src/stringx.cpp
@@ -206,7 +206,7 @@ void setFormulaWithString(string formulastring, Formula* formula, Module* module
         postnum = input.tellg();
       }
       string num;
-      num.assign(formulastring, prenum, static_cast<size_t>(postnum-prenum));
+      num.assign(formulastring, static_cast<size_t>(prenum), static_cast<size_t>(postnum-prenum));
       //cout << "Text: '" << num << "'" << endl;
       //cout << "Number: '" << number << "'" << endl;
       formula->AddText(&num);
